requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.3
dj-database-url==2.2.0
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
exponent_server_sdk==2.1.0
fillpdf==0.7.3
google-auth==2.35.0
gunicorn==22.0.0
idna==3.10
kombu==5.4.2
packaging==24.0
pdf2image==1.17.0
pdfrw2==0.5.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
